requirements.txt
Flask==2.3.2
Jinja2==3.1.2
mysql-connector-python